Demo commit : As code has error

+++ opd-appointment-service/src/main/java/com/ztrios/opd_appointment_service/dto/AppointmentResponse.java
@@ -15,6 +15,6 @@ public record AppointmentResponse(UUID appointmentId,
 //                                  Instant createdAt,
 
                                   AppointmentStatus status,
-                                  Long serialNo,
+                                  Integer serialNo,
                                   String paymentLink ) {
 }
